refactor(bfloat16): split bfloat16 into core / manipulators / iostream (#1334)
First type of Phase 2 group 5a (#1444). bfloat16 gets a core.hpp that pulls none
of the five I/O-family headers, so a compute kernel does not pay for text
formatting it never calls.

    core.hpp          44,559 lines / 247 headers / 0 I/O-family
    bfloat16.hpp      63,751 lines / 341 headers / 5 I/O-family  (was 93,035 / 377 / 5)

The umbrella keeps its name and still includes every layer, so no caller changes.

What moved:

- parse(), to_binary() and to_native() build strings through a stringstream and
  are now in manipulators.hpp. parse() was already declared in bfloat16_fwd.hpp.
- operator<< and operator>> are now in iostream.hpp. The dependency runs
  iostream -> manipulators here (operator>> hands its token to parse), and
  manipulators does not include iostream back, so there is no cycle.
- assign(const std::string&) stays in the core: it walks a "0b...." pattern with
  plain std::string operations. Its seven std::cerr diagnostics became
  fprintf(stderr, ...), the Phase 0 idiom, which is what lets the core drop
  <iostream>.
- BFLOAT_ENABLE_LITERALS and BFLOAT_THROW_ARITHMETIC_EXCEPTION now default in
  bfloat16_impl.hpp beside the code they govern, so the core-only path gets the
  same defaults the umbrella used to supply (#1436).

Two incidental finds, both include hygiene rather than layering:

- bfloat16_impl.hpp included <regex> and used none of it. Removing it is most of
  the umbrella's 93,035 -> 63,751 drop; the split itself accounts for the rest.
- attributes.hpp used std::setw without <iomanip> and reached type_tag and
  operator<< through the umbrella's include order. It now names them.

isinteger() called sw::universal::floor(bfloat16), which lives in mathlib and so
is not in the core -- an undefined-inline warning and a link error waiting for
the first core-only caller. That overload is bfloat16(std::floor(float(x))) and
bfloat16 -> float is exact, so it is inlined as std::floor(f) == f. Same
predicate, bit for bit.

Verification:

- Layer gate: a TU including core.hpp alone, and one including
  core.hpp + manipulators.hpp, both compile, link and run -- gcc 13.3 and clang.
- 13 bfloat16 test targets plus 14 downstream consumers that link the type
  (elreal EFT/ZBCL, multifile, the fma trait tests, cfloat_bfloat16): 27 targets
  built and run green on both compilers.
- Differential against main over all 65,536 encodings through to_binary,
  to_native, to_hex, hex_print, to_triple, color_print, operator<<, sign, scale,
  significant and isinteger, plus parse/operator>> over 22 decimal texts and
  assign() over every diagnostic path: 26,413,230 bytes of stdout and all seven
  stderr diagnostics byte-identical, on gcc and clang.

Found, not fixed: type_field(bfloat16) reads BfloatType::fbits, which bfloat16
does not define. It is broken identically on main -- the template is simply never
instantiated anywhere -- so it is not this PR's to change.
