+++ include/sw/universal/number/bfloat16/bfloat16.hpp
@@ -20,21 +20,11 @@
 
 ////////////////////////////////////////////////////////////////////////////////////////
 ///  BEHAVIORAL COMPILATION SWITCHES
-
-////////////////////////////////////////////////////////////////////////////////////////
-// enable/disable the ability to use literals in binary logic and arithmetic operators
-#if !defined(BFLOAT_ENABLE_LITERALS)
-// default is to enable them
-#define BFLOAT_ENABLE_LITERALS 1
-#endif
-
-////////////////////////////////////////////////////////////////////////////////////////
-// enable throwing specific exceptions for integer arithmetic errors
-// left to application to enable
-#if !defined(BFLOAT_THROW_ARITHMETIC_EXCEPTION)
-// default is to use std::cerr for signalling an error
-#define BFLOAT_THROW_ARITHMETIC_EXCEPTION 0
-#endif
+///
+/// BFLOAT_ENABLE_LITERALS and BFLOAT_THROW_ARITHMETIC_EXCEPTION now default in
+/// bfloat16_impl.hpp, beside the code they govern, so that a translation unit which
+/// includes core.hpp directly gets the same defaults this umbrella used to supply
+/// (#1334, #1436). Defining either before this header still wins, as before.
 
 ///////////////////////////////////////////////////////////////////////////////////////
 // bring in the trait functions
@@ -44,14 +34,16 @@
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// INCLUDE FILES that make up the library
-#include <universal/number/bfloat16/exceptions.hpp>
-#include <universal/number/bfloat16/bfloat16_impl.hpp>
-#include <universal/traits/bfloat16_traits.hpp>
-#include <universal/number/bfloat16/numeric_limits.hpp>
+// layer 1: the arithmetic core (#1334). Include core.hpp directly in a translation
+// unit that only computes -- it pulls no <iostream>/<sstream>/<iomanip>.
+#include <universal/number/bfloat16/core.hpp>
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// useful functions to work with bfloats
+// layer 2a: the string producers -- parse, to_binary, to_hex, to_triple, color_print
 #include <universal/number/bfloat16/manipulators.hpp>
+// layer 2b: the <iostream> half -- operator<< / operator>>
+#include <universal/number/bfloat16/iostream.hpp>
 #include <universal/number/bfloat16/attributes.hpp>
 
 ///////////////////////////////////////////////////////////////////////////////////////
